+++ packages/mds-agency/api.ts
@@ -17,9 +17,9 @@
 import express from 'express'
 
 import log from '@mds-core/mds-logger'
-import { pathsFor, providerClaimMiddleware } from '@mds-core/mds-utils'
+import { pathsFor } from '@mds-core/mds-utils'
 import { AgencyApiRequest, AgencyApiResponse } from '@mds-core/mds-agency/types'
-import { checkAccess } from '@mds-core/mds-api-server'
+import { checkAccess, providerClaimMiddleware } from '@mds-core/mds-api-server'
 import {
   getAllServiceAreas,
   getServiceAreaById,
