+++ packages/mds-api-server/tests/index.spec.ts
@@ -16,9 +16,13 @@
     limitations under the License.
  */
 
+import assert from 'assert'
 import supertest from 'supertest'
 import test from 'unit.js'
-import { ApiServer } from '../index'
+import Sinon from 'sinon'
+import uuid from 'uuid'
+import { TEST1_PROVIDER_ID } from '@mds-core/mds-providers'
+import { ApiServer, ApiRequest, providerClaimMiddleware, ProviderClaimResponse } from '../index'
 
 const request = supertest(ApiServer(app => app))
 
@@ -117,5 +121,93 @@ describe('Testing API Server', () => {
       .end(err => {
         done(err)
       })
+  })
+
+  describe('Provider claims helper', () => {
+    it('spits out missing claims', async () => {
+      const req: ApiRequest = {} as ApiRequest
+
+      const send = Sinon.fake.returns('boop')
+      const status = Sinon.fake.returns({ send })
+      const res: ProviderClaimResponse = ({
+        status,
+        locals: {
+          claims: null
+        }
+      } as unknown) as ProviderClaimResponse
+
+      const result = await providerClaimMiddleware(req, res)
+
+      assert.strictEqual(status.calledOnceWithExactly(400), true)
+      assert.strictEqual(send.calledOnce, true)
+      assert.strictEqual(result, false)
+
+      Sinon.restore()
+    })
+
+    it('spits out bad uuid', async () => {
+      const req: ApiRequest = {} as ApiRequest
+
+      const send = Sinon.fake.returns('boop')
+      const status = Sinon.fake.returns({ send })
+      const res: ProviderClaimResponse = ({
+        status,
+        locals: {
+          claims: {
+            provider_id: 'not-a-uuid'
+          }
+        }
+      } as unknown) as ProviderClaimResponse
+
+      const result = await providerClaimMiddleware(req, res)
+
+      assert.strictEqual(status.calledOnceWithExactly(400), true)
+      assert.strictEqual(send.calledOnce, true)
+      assert.strictEqual(result, false)
+
+      Sinon.restore()
+    })
+
+    it('spits out non provider uuid', async () => {
+      const req: ApiRequest = {} as ApiRequest
+
+      const send = Sinon.fake.returns('boop')
+      const status = Sinon.fake.returns({ send })
+      const res: ProviderClaimResponse = ({
+        status,
+        locals: {
+          claims: {
+            provider_id: uuid()
+          }
+        }
+      } as unknown) as ProviderClaimResponse
+
+      const result = await providerClaimMiddleware(req, res)
+
+      assert.strictEqual(status.calledOnceWithExactly(400), true)
+      assert.strictEqual(send.calledOnce, true)
+      assert.strictEqual(result, false)
+
+      Sinon.restore()
+    })
+
+    it('allows legit provider id', async () => {
+      const req: ApiRequest = {} as ApiRequest
+      const provider_id = TEST1_PROVIDER_ID
+      const send = Sinon.fake.returns('boop')
+      const status = Sinon.fake.returns({ send })
+      const res: ProviderClaimResponse = ({
+        status,
+        locals: {
+          claims: { provider_id }
+        }
+      } as unknown) as ProviderClaimResponse
+
+      const result = await providerClaimMiddleware(req, res)
+
+      assert.strictEqual(result, provider_id)
+
+      Sinon.restore()
+    })
   })
 })
